Add Widgets Final2

+++ app.py
@@ -12,7 +12,7 @@ if 'df_experiment_results' not in st.session_state:
 
 st.header('Tossing a Coin')
 
-chart = st.line_chart([0.5])
+chart = st.line_chart([0.5]).astype(float)
 
 def toss_coin(n):
